# Remove jquery