1. Remove all jQuery dependencies.
2. Remove Promise and async methods.
3. Only keep core feature for sanitize and remove others like insert result to document.