@randomdross 2016-02-03
This is fantastic, thank you!  At a high level though I do have an issue with making jSanity synchronous...  jSanity is async so that it can break up processing into small chunks and not block page rendering on the sanitize call.  This is a pretty important feature IMO, since otherwise sanitization can sit there completely blocking the UI thread for seconds (assuming a large chunk of input HTML to sanitize).  I would have liked to do everything in a web worker, but unfortunately web workers prohibit access to the DOM.