@randomdross 2016-02-03
...that being said, removing the dependency on jQuery is excellent!  And the global namespace is good too.