@kh9n 2016-02-03
@randomdross Thanks David! That is a great valid point about the performance concern. However, I think current async version has exactly same performance drop when a really huge HTML code need to be sanitize, that is the core logic (**treeWalk**) navigate inside of dom for purify is a recursive and synchronize method -- it will not return until all node from given HTML be processed.

to simulate the case for huge or complex HTML code, I force a sleep at entry of "treeWalk" for seconds, even for current async implementation, it will still hang the UI. You can find the demo here: [](http://jsbin.com/vipuzagoha/edit?html,js,console,output)

So, I think the better way to mitigate the perf concern on huge HTML is make the treeWalk to be a-synchronized.