@randomdross 2016-02-03
If you're seeing it hang like that it would be a bug.  I'm seeing the correct behavior (no blocking) on the benchmark page.  My process:
1)  Load the benchmark page: http://jsanity.azurewebsites.net/jsanity-benchmark-pretty.htm
(I'm using Chrome, though you should see the same result in FF or Edge.)
2)  Click the "Sanitize with jSanity" button.
3)  Observe "synchronous return" after 1ms, whereas all processing takes a second or so.