@randomdross 2016-02-03
If you just sleep for 1 second, anywhere (outside of a web worker), this will always hang the browser.  Async code can't help with that.  The only thing that would prevent that is if the code were truly multi-threaded and used a web worker.  Code that uses the DOM can't do that, unfortunately.  

The async approach enables cooperative multi-tasking, so code on the page has an opportunity to execute in between batches of sanitization workload.

Try this:
1) Load the benchmark page: http://jsanity.azurewebsites.net/jsanity-benchmark-pretty.htm
(I'm using Chrome, though you should see the same result in FF or Edge.)
2) Click the "Sanitize with jSanity" button.
3) Observe "synchronous return" after 1ms, whereas all processing takes a second or so.
I should probably add an animation or something on the benchmark page so it's easier to see how code on the page isn't blocked waiting on the entire sanitization workload to complete before it can execute.