@kh9n 2016-02-04
@randomdross I was trying to load a complex HTML page through ajax rather than use sleep, but it can't make $.ajax work on jsbin. :(

For the story of use async or sync, I think we still could keep the sanitize process as synchronize (as the core logic right now is synchronize), but let user make their best decision. 

Like, people still able to wrap the synchronize call into any promise implementation or just native setTimeout manner, then run those in the async way.
For people who don't have performance concern, they can choose the synchronize way, or add asynchronize wrapper base on their needs.

My point of view is, as support async run is not the core functionality of jSanity (at least, not the priority for now), by removing that restriction, we could give user some flexibility and enable more scenarios.