@randomdross 2016-02-04
If the API is synchronous, the best that a caller can do with a 5 second sanitization operation is to call the API and have it block for a full five seconds.  There's no opportunity during those 5 seconds for other script to run.
For many applications I would expect it to be a non-starter to have an API that blocks on the main thread like that.
I wouldn't be opposed to a flag to call the API synchronously, but I do think it does need to maintain its async ability.  Or otherwise find some magical way to use web workers.
It would be amazing to be able to get rid of jQuery though.  This becomes more feasible as older browsers go out of support and we can move to ES6 promises.

Date: Wed, 3 Feb 2016 22:49:55 -0800
From: notifications@github.com
To: JSanity@noreply.github.com
CC: dross@pobox.com
Subject: Re: [JSanity] Remove jquery (#10)

@randomdross I was trying to load a complex HTML page through ajax rather than use sleep, but it can't make $.ajax work on jsbin. :(


For the story of use async or sync, I think we still could keep the sanitize process as synchronize (as the core logic right now is synchronize), but let user make their best decision. 


Like, people still able to wrap the synchronize call into any promise implementation or just native setTimeout manner, then run those in the async way.

For people who don't have performance concern, they can choose the synchronize way, or add asynchronize wrapper base on their needs.


My point of view is, as support async run is not the core functionality of jSanity (at least, not the priority for now), by removing that restriction, we could give user some flexibility and enable more scenarios.


—
Reply to this email directly or view it on GitHub.