@kh9n 2016-02-05
Some outline:
1. Open the door both for async or sync, it will auto switch depends on if any onFinishedCallback passed in.
2. Try to address the concern for the synchronized way of treeWalker for huge HTML, basically I wrap each node processing into a single job, and those jobs could be run continuously(sync), or in parallel(async).

for "huge" HTML (I use https://github.com/Microsoft/JSanity/blob/master/jsanity-0.2.js as test case):
in sync mode: it takes around 280 ms
in async mode: it takes around 30 seconds
in optimized async mode: it takes around 330 ms