@randomdross 2016-02-06
Oh wow, that looks fantastic!  Give me a few days to play around with this a bit before I make the decision to pull it in, it's a big change.
Are you saying that async was originally 30 seconds but then you optimized and got it down to 330ms?  (Or are there actually two different async modes?)
Date: Fri, 5 Feb 2016 04:35:41 -0800
From: notifications@github.com
To: JSanity@noreply.github.com
CC: dross@pobox.com
Subject: Re: [JSanity] Remove jquery (#10)

Some outline:

1. Open the door both for async or sync, it will auto switch depends on if any onFinishedCallback passed in.

2. Try to address the concern for the synchronized way of treeWalker for huge HTML, basically I wrap each node processing into a single job, and those jobs could be run continuously(sync), or in parallel(async).


for "huge" HTML (I use https://github.com/Microsoft/JSanity/blob/master/jsanity-0.2.js as test case):

in sync mode: it takes around 280 ms

in async mode: it takes around 30 seconds

in optimized async mode: it takes around 330 ms


—
Reply to this email directly or view it on GitHub.