@randomdross 2016-02-11
Sorry for the delay!  Having spent some time with the new code, I'm now even more excited about it!  Take a look at this feedback and make any changes, then I'll merge your version into the original.

Ok, here's a version of my original benchmark page that's updated to use your implementation:
http://jsanity.azurewebsites.net/jsanity-benchmark-pretty-test.htm

For comparison, here's the benchmark page for the original implementation:
http://jsanity.azurewebsites.net/jsanity-benchmark-pretty.htm

Just to be consistent, both my benchmark pages insert the sanitized nodes into the DOM.

Notes:
  - On my machine, the original jSanity returns synchronously immediately (3ms).  In async mode, the new code takes a little longer (~150ms).  This isn't a big deal, but I wonder why it isn't returning more immediately.
  - The original jSanity finishes all its work in ~950ms on average.
  - The new jSanity finishes in ~1000ms on average.  However it seems that the new implementation yields quite a bit more so other activities on the page will have more of a chance to operate.  Here's what I see:
Original run:   XXXXXXXXXXXXXX.XXX.XXX.#
Your new code:  XXXX.XX..............XX.............XX..#

  - The new version (synchronous) finishes in ~850ms.
New code, sync: XXXXXXXXXXXXXXXXX..#

Other notes:
  - I noticed that you're missing the support for maxWidth / maxHeight / overflow enforcement.  Can you put it back?  =)  This is a security feature.
    - The code is at line 870 in https://github.com/Microsoft/JSanity/blob/master/jsanity-0.2.js
  - I like that you just return a node that can be appended as necessary.  That makes more sense than the way I was doing it.
  - I seem to have only half-removed the support for toStaticHTML from my code.  It looks like you re-implemented the part I removed.  Thanks!  =)  We can probably remove the old IE compat code at this point since IE < 11 is now unsupported.